requirements.txt
fastapi==0.68.2
pydantic==1.10.8
uvicorn==0.15.0
httpx==0.24.1
python-gvm==23.5.1
beautifulsoup4==4.12.2
## The following requirements were added by pip freeze:
aiofiles==22.1.0
aiosqlite==0.19.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
attrs==21.4.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
bleach==6.0.0
cairocffi==1.5.1
CairoSVG==2.7.0
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colour==0.1.5
comm==0.1.3
commonmark==0.9.1
cpe==1.2.1
cryptography==41.0.1
cssselect2==0.7.0
cybox==2.1.0.21
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
drawSvg==1.9.0
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
fire==0.4.0
fqdn==1.5.1
h11==0.14.0
httpcore==0.17.2
idna==3.4
imageio==2.31.0
importlib-metadata==3.3.0
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
loguru==0.7.0
lxml==4.9.2
maec==4.1.0.17
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
mitreattack-python==2.0.11
mixbox==1.0.5
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
netaddr==0.8.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
paramiko==3.2.0
parso==0.8.3
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
pluralizer==1.2.0
pooch==1.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pyattck==7.1.2
pyattck-data==2.6.3
pycountry==22.3.5
pycparser==2.21
Pygments==2.15.1
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
requests==2.30.0
requests-cache==1.0.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
Send2Trash==1.8.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.14.2
stix==1.2.0.11
stix2==3.0.1
stix2-elevator==4.1.7
stix2-patterns==2.0.0
stix2-validator==3.1.3
stixmarx==1.0.8
tabulate==0.9.0
taxii2-client==2.3.0
websocket-client==1.5.1
win32-setctime==1.1.0
XlsxWriter==3.1.2
xmltodict==0.13.0
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.15.0